adds metadata remapping
Defines the mapping for the metadata defined in p4d2model to the bmv2
simple switch intrinsic/standard metadata and provides a read function
for the file that returns the remapping (map of string to string).

We need to hook this up with p4c such that it knows which file to load
depending on the architecture and the target. And probably just simply
visit all the metadata declarations in the program and add a name
annotation to remap the name. I hope that fields can take qualified
names.