adds field_aliases to support remapping of metadata names
BMV2 supports field_aliases to facilitate using different names
in the P4 programs from the names supported by the target.
For PSA and P4D2 we define a different set of metadata and a map
file to map to bmv2 simple_switch (aka v1model). The name map name
is currently hardcoded, and it should be set by the compiler driver
as it knows which architecture and which target.

Also removes the force_arith field in json, since it is no longer used.

This started initially as a pass to rename metadata in the program.
It seems to not be necessary, so the code is #ifdef-ed out. We should
remove it in the final version if we find that it is really not
needed.