expands the v1model standard metadata to cover all metadata exposed by simple switch
Allows access to bmv2 simple switch intrinsic and queueing metadata
for programs that use the v1model. While these metadata would be
available by defining separate metadata with the correct names, it is
desirable that the functionality is exposed in the architecture
definition. This patch will work with any metadata that has fields
renamed.